Change for testing on Iota
This should be temporary changes in ham.c for logging externally to
iota's headnode. This external logging target should really be specified
in an external configuration file or somehow in the topology management
file, if the feature is kept. This is just a temporary fix so that some
metrics can be gathered for ADDAM running on the iota cluster.